+++ src/primitives/components/index.tsx
@@ -4,6 +4,7 @@ export * from "./classes.js";
 export * from "./controls.js";
 export * from "./display.js";
 export * from "./key_value.js";
+export * from "./link.js";
 export * from "./layout.js";
 export * from "./shared.js";
 export * from "./time_counter.js";